requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
netmiko==5.0.0
schedule==1.1.0
pyyaml==6.0